+++ src/btree/bt_read.c
@@ -96,7 +96,7 @@ __page_read(WT_SESSION_IMPL *session, WT_REF *ref, uint32_t flags)
     WT_ITEM tmp;
     WT_PAGE *notused;
     uint32_t page_flags;
-    uint8_t previous_state;
+    uint32_t previous_state;
     bool prepare;
 
     /*
@@ -267,7 +267,7 @@ __wt_page_in_func(WT_SESSION_IMPL *session, WT_REF *ref, uint32_t flags
     WT_PAGE *page;
     WT_TXN *txn;
     uint64_t sleep_usecs, yield_cnt;
-    uint8_t current_state;
+    uint32_t current_state;
     int force_attempts;
     bool busy, cache_work, evict_skip, stalled, wont_need;
