+++ src/btree/bt_sync_obsolete.c
@@ -190,7 +190,7 @@ static int
 __sync_obsolete_cleanup_one(WT_SESSION_IMPL *session, WT_REF *ref)
 {
     WT_DECL_RET;
-    uint8_t new_state, previous_state;
+    uint32_t new_state, previous_state;
     bool busy, ref_deleted;
 
     busy = ref_deleted = false;
